day11-工作台相关接口

+++ df-pojo/src/main/java/com/dineflow/vo/BusinessDataVO.java
@@ -6,6 +6,7 @@ import lombok.Data;
 import lombok.NoArgsConstructor;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
 
 /**
  * 数据概览
@@ -16,14 +17,14 @@ import java.io.Serializable;
 @AllArgsConstructor
 public class BusinessDataVO implements Serializable {
 
-    private Double turnover;//营业额
+    private BigDecimal turnover;//营业额
 
-    private Integer validOrderCount;//有效订单数
+    private Long validOrderCount;//有效订单数
 
-    private Double orderCompletionRate;//订单完成率
+    private BigDecimal orderCompletionRate;//订单完成率
 
-    private Double unitPrice;//平均客单价
+    private BigDecimal unitPrice;//平均客单价
 
-    private Integer newUsers;//新增用户数
+    private Long newUsers;//新增用户数
 
 }
